fix: segmented code and add delay into while loop
